+++ tests/Unit/AcDeviceTest.php
@@ -10,13 +10,21 @@ class AcDeviceTest extends TestCase
 {
     public function test_creating_ac_device_from_connectlife_api()
     {
-        $acDevice = new AcDevice(
-            File::json(base_path('/tests/hijuconn-api-data/device-status.json'))[0],
-            File::json(base_path('/tests/hijuconn-api-data/devices-config.json'))['117'],
+        putenv(
+            'DEVICES_CONFIG=' .
+            File::get(base_path("/tests/hijuconn-api-data/devices-config.json"))
         );
 
-        $this->assertGreaterThan(0, count($acDevice->fanSpeedOptions));
-        $this->assertIsArray($acDevice->toHomeAssistantDiscoveryArray());
-        $this->assertIsArray($acDevice->toConnectLifeApiPropertiesArray());
+        $deviceFeatureCodes = ['117', '104', '109'];
+
+        foreach ($deviceFeatureCodes as $deviceFeatureCode) {
+            $acDevice = new AcDevice(
+                File::json(base_path("/tests/hijuconn-api-data/device-status-$deviceFeatureCode.json"))[0]
+            );
+
+            $this->assertGreaterThan(0, count($acDevice->fanSpeedOptions));
+            $this->assertIsArray($acDevice->toHomeAssistantDiscoveryArray());
+            $this->assertIsArray($acDevice->toConnectLifeApiPropertiesArray());
+        }
     }
 }