+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2023-02-25
+
+### Added
+
+### Fixed
+
+- Crash when swing is not supported.
+- "power" on/off command.
+- Using "deviceFeatureCode" instead of "deviceFeatureCode" to differentiate AC devices model.
+
 ## [2.0.0] - 2023-02-24
 
 ### Added